// web/src/api/models/TypesBGPPeerSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesBGPPeerSummary
 */
export interface TypesBGPPeerSummary {
    /**
     * 
     * @type {string}
     * @memberof TypesBGPPeerSummary
     */
    address: string;
    /**
     * 
     * @type {number}
     * @memberof TypesBGPPeerSummary
     */
    asn: number;
    /**
     * 
     * @type {string}
     * @memberof TypesBGPPeerSummary
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof TypesBGPPeerSummary
     */
    msg_rcvd: number;
    /**
     * 
     * @type {number}
     * @memberof TypesBGPPeerSummary
     */
    msg_sent: number;
    /**
     * 
     * @type {number}
     * @memberof TypesBGPPeerSummary
     */
    prefixes: number;
    /**
     * 
     * @type {string}
     * @memberof TypesBGPPeerSummary
     */
    state: string;
    /**
     * 
     * @type {string}
     * @memberof TypesBGPPeerSummary
     */
    uptime: string;
}

/**
 * Check if a given object implements the TypesBGPPeerSummary interface.
 */
export function instanceOfTypesBGPPeerSummary(value: object): value is TypesBGPPeerSummary {
    if (!('address' in value) || value['address'] === undefined) return false;
    if (!('asn' in value) || value['asn'] === undefined) return false;
    if (!('msg_rcvd' in value) || value['msg_rcvd'] === undefined) return false;
    if (!('msg_sent' in value) || value['msg_sent'] === undefined) return false;
    if (!('prefixes' in value) || value['prefixes'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('uptime' in value) || value['uptime'] === undefined) return false;
    return true;
}

export function TypesBGPPeerSummaryFromJSON(json: any): TypesBGPPeerSummary {
    return TypesBGPPeerSummaryFromJSONTyped(json, false);
}

export function TypesBGPPeerSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesBGPPeerSummary {
    if (json == null) {
        return json;
    }
    return {
        
        'address': json['address'],
        'asn': json['asn'],
        'description': json['description'] == null ? undefined : json['description'],
        'msg_rcvd': json['msg_rcvd'],
        'msg_sent': json['msg_sent'],
        'prefixes': json['prefixes'],
        'state': json['state'],
        'uptime': json['uptime'],
    };
}

export function TypesBGPPeerSummaryToJSON(json: any): TypesBGPPeerSummary {
    return TypesBGPPeerSummaryToJSONTyped(json, false);
}

export function TypesBGPPeerSummaryToJSONTyped(value?: TypesBGPPeerSummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'address': value['address'],
        'asn': value['asn'],
        'description': value['description'],
        'msg_rcvd': value['msg_rcvd'],
        'msg_sent': value['msg_sent'],
        'prefixes': value['prefixes'],
        'state': value['state'],
        'uptime': value['uptime'],
    };
}